Fixed setup.py

+++ setup.py
@@ -1,7 +1,8 @@
 from commands import server
-from commands import rank,messages
+from commands import rank, messages
 from core.CRUD import db
 from psycopg2.errors import DuplicateTable
+from os.path import exists
 
 try:
     server.CreateServerTable()
@@ -11,19 +12,19 @@ try:
 except DuplicateTable:
     print("Tables already exists")
     print("TYPE 'reset' IN ORDER TO DELETE THE OLD TABLES AND CREATE NEW ONES. BE CAREFULL, THIS WILL ERASE ALL DATA")
-    
-    if(input().lower() == 'reset'):    
+
+    if(input().lower() == 'reset'):
         db.custom_insert("DROP TABLE ServerConfigs")
         db.custom_insert("DROP TABLE Rank")
         db.custom_insert("DROP TABLE Messages")
-        
+
         server.CreateServerTable()
         rank.CreateRankTable()
         messages.CreateMessagesTable()
-
-f = open("localToken.txt", "w")
-f.write("Remove this and paste your bot token here")
-f.close()
+if not exists("localToken.txt"):
+    f = open("localToken.txt", "w")
+    f.write("Remove this and paste your bot token here")
+    f.close()
 f = open("immoral_chars.json", "w")
 f.write("""{
     "𝓘" : "I",
